SkillAPI v2.4

+++ SkillAPI/src/com/sucy/skill/api/event/PlayerOnDamagedEvent.java
@@ -16,7 +16,7 @@ public class PlayerOnDamagedEvent extends Event {
     private Hashtable<String, Object> flags = new Hashtable<String, Object>();
     private Player player;
     private LivingEntity attacker;
-    private int damage;
+    private double damage;
 
     /**
      * Constructor - you shouldn't need to use this
@@ -25,7 +25,7 @@ public class PlayerOnDamagedEvent extends Event {
      * @param entity mob or player damaging the player
      * @param damage amount of damage dealt
      */
-    public PlayerOnDamagedEvent(Player player, LivingEntity entity, int damage) {
+    public PlayerOnDamagedEvent(Player player, LivingEntity entity, double damage) {
         this.player = player;
         this.attacker = entity;
         this.damage = damage;
@@ -48,7 +48,7 @@ public class PlayerOnDamagedEvent extends Event {
     /**
      * @return the damage dealt
      */
-    public int getDamage() {
+    public double getDamage() {
         return damage;
     }
 
@@ -57,7 +57,7 @@ public class PlayerOnDamagedEvent extends Event {
      *
      * @param amount new amount of damage
      */
-    public void setDamage(int amount) {
+    public void setDamage(double amount) {
         damage = amount;
     }
